requirements.txt
datasets>=2.16.0
pandas>=2.0.0
pyarrow>=14.0.0
pytest>=7.4.0
groq>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
